I AM COMITTING THE BETA OF THE HOME PAGE
Currently HTML has a "home page section". This simply
holds an image of a Tic Tic Toe board, and has two buttons
to either create an account or to sign in to an existing
account. When you click either button, an alert opens stating that
the feature is not implemented.

There is some very basic styling using SCSS, Flexbox, and Boostrap.
The styling of the home page is one of the last things that I shall do.