I AM COMMITING A BETA OF THE SIGNIN PAGE
This page is just a beta. It makes a call to the sign-in web service.
It then updates the status message area with a success or failure
message. It stores the authentication token in the store.

I also did some basic, boring, bootstrap styling that looks like
every other website. Styling will be my last task.