I AM COMMITING A VERY ALPHA VERSION OF NEW GAME
When the user clicks the "New Game" button in the "Game Options" page,
the game board now appears. The call to the web service to create a
new game works, and the game is attached to the store object.

The game board itself is very alpha. It uses a Bootstrap grid, with
basic "X" text elements in each grid cell. There is a return to the
"Game Options" screen button that is always enabled.

This is good enough for the moment. I can start working on the game
play, but I want to stop and cleanup, then I want to make a state
machine. There is too much repetition of hiding and showing things,
and that should be handled in one place.

I also have a working curl script for creating a new game.