THIS COMMITS CONCERNS THE README.MD
I renamed all of the wireframe bits in order to remove whitespace from
the file names.

I have a beta version of the README.MD file.

I also found an issue with my curl script for updating a game. It
had a type in it. But I decided to use Postman instead. I really don'
like curl.