THE GAMBOARD NOW USES GRAPHICS FOR THE X AND O CHARACTERS
Paul made me SVG and PNG images that I now use for the gameboard. I
had a little difficulty because it appears that when hide and then
show something with an image that the image doesn't load into the
HTML. So I load the X and O image and hide them. I then use jQuery to
clone them then appened them to the game board.