I AM WORKING ON UPDATE GAME
I tried to invoke the update web service for a game from the game
engine. But I realized that doing so creates a circular reference
with respect to CommonJS modules.

But I do have the rest of the code in place. I want to make a commit
now.

I also improved the game play. The scoring now checks that we are on
move 5 or later in order to win. I also simplified some code.