THE GAME NOW USES THE UPDATE GAME API
All that is essentially left is to show the number of games played.

I had to refactor my method for scoring into two methods. That is
because of the circular references to gameScoringEngine. I had to add
a method that returns the game board from the web service to then update
the gameboard.

The board now bounces when playing a game. I need to fix that in a
future commit.