I FIXED A BUG I FOUND IN GAME PLAY
My image was scaled to be small. When testing that you cannot select
a cell already played, I noticed that the correct thing happens When
I click on the cell backgound DIV. But if I click on the image itself,
Then I receive an "undefined" in the status message. I not check for
a click on either the gamepiece image, or the Bootstrap grid cell
directly. Both indicate that I cannot play the same cell again.