I ADDED AN EDGE CASE TO THE GAME gameScoringEngine
I realized that through naive game play, X can win on the last move by
placing X in one of the corners and winning by having Xs in either
diagonal, as well as having having X's in the top or bottom row, or the
left or right columns. My scoring algorithm will detect that.

I made small tweaks to README.md including my edge case in the previous
paragrah.

When I deployed yesterday to GitHub, my images were not loading. I
altered the path to the image on the home page to see if that fixes
the deployed game. Regardless, I have a ticket open with the
instructors.